DrinkMakerWebApp/Backend/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyserial
psutil
python-multipart